requirements.txt
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
boto3==1.42.35
botocore==1.42.35
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.3
distro==1.9.0
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
humanfriendly==10.0
idna==3.11
jiter==0.12.0
jmespath==1.1.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
openai==2.15.0
packaging==26.0
protobuf==6.33.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
fastapi==0.111.0
uvicorn==0.30.1
requests==2.31.0
